Add new cop `Lint/UselessTimes`.
This cop looks for pointless uses of `Integer#times` and replaces or removes them as necessary. When the integer is <= 0, the block will never be yielded to, so the entire thing can be removed. When the integer is 1, the block will only ever yield once, so the `1.times` call can be replaced with the block body.