Add new Lint/ConstantDefinitionInBlock cop
Constants defined within a block are added to the current lexical scope,
which means the block has no effect on where the constant is defined or
how it can be accessed. I often see constants defined inside Rake tasks
or RSpec `describe` blocks where the intent was clearly for them to be
scoped to the block, but which actually leak into the top-level scope.